% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_style.R
\name{apa_style}
\alias{apa_style}
\title{APA-Like Table Style}
\usage{
apa_style(table)
}
\arguments{
\item{table}{A \code{gt} table}
}
\value{
A formatted \code{gt} table
}
\description{
Style a \code{gt} table minimally with an APA-like style.
}
\examples{
library(gt)

mtcars \%>\%
gt() \%>\%
apa_style()
}
