% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bene_project.R
\name{bene_project}
\alias{bene_project}
\title{Create a New Project}
\usage{
bene_project(path, ...)
}
\arguments{
\item{path}{Filepath where project should be created}

\item{...}{Optional arguments}
}
\value{
A file structure
}
\description{
This function creates a new project with the correct folders and some files,
that are always used.
}
