% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_summaries.R
\name{mean_sd}
\alias{mean_sd}
\title{Compute Mean and its Standard Deviation for ggplot2}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A tibble with columns \code{ymin}, \code{mean}, and \code{ymax}.
}
\description{
Compute mean and standard deviation for a given vector. Can be used as
\code{fun.data} in ggplot2's \code{\link[ggplot2]{stat_summary}}. This comes
in very handy for plotting means and standard deviations as no further
arguments must be submitted.
}
\examples{
x <- rnorm(200)

mean_sd(x)
}
