% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_summaries.R
\name{median_se}
\alias{median_se}
\title{Compute Median and its Standard Error for ggplot2}
\usage{
median_se(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A tibble with columns \code{ymin}, \code{mean}, and \code{ymax}.
}
\description{
Compute median and its standard error for a given vector. Can be used as
\code{fun.data} in ggplot2's \code{\link[ggplot2]{stat_summary}}.
}
\examples{
x <- rnorm(200)

median_se(x)
}
