% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_working_directory.R
\name{open_working_directory}
\alias{open_working_directory}
\title{Open Current Working Directory}
\usage{
open_working_directory()
}
\description{
Shortcut to open the current working directory from the R console. This is
useful if one uses a file-launcher like one in Microsoft PowerToys and opens
the project file instead of the actual project folder.
}
