% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{palmer_penguins}
\alias{palmer_penguins}
\title{Palmer Penguin Dataset}
\format{
A tibble with 333 rows and 8 columns
\describe{
\item{species}{Penguin's species}
\item{island}{Penguin's island}
\item{bill_length_mm}{Bill length in mm}
\item{bill_depth_mm}{Bill depth in mm}
\item{flipper_length_mm}{Flipper length in mm}
\item{body_mass_g}{Weight in g}
\item{sex}{Sex}
\item{year}{Year of observation}
}
}
\source{
\url{https://github.com/allisonhorst/palmerpenguins}
}
\usage{
palmer_penguins
}
\description{
A cleaned version of the Palmer Penguins dataset from the package
\code{palmerpenguins} (i.e. all missing removed). This serves primarily as an
example dataset for plotting examples.
}
\keyword{datasets}
