% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_custom_plot.R
\name{save_custom_plot}
\alias{save_custom_plot}
\title{Save A Plot With Custom Theming}
\usage{
save_custom_plot(
  plot,
  filename,
  path,
  width = 16,
  height = 16,
  units = "cm",
  dpi = 320
)
}
\arguments{
\item{plot}{A ggplot2 plot.}

\item{filename}{The plot's desired file name.}

\item{path}{The desired path to which the plot should be saved.}

\item{width}{Plot width, defaults to 16.}

\item{height}{Plot height, defaults to 16.}

\item{units}{Plot dimension unit, defaults to "cm".}

\item{dpi}{Plot resolution, defaults to 320.}
}
\value{
Nothing, saves plot to disk.
}
\description{
The various \verb{use_*} functions from the benelib are powered by the thematic
package. This powerful package makes it possible to use custom Google fonts
without worrying about plotting those. However, this comes with a small
drawback for saving plots. If one uses ggplot2's \code{ggsave} function, text may
be rendered too small in raster images. \code{thematic} provides a custom
function, for which this function is a simple wrapper.
}
\details{
In short: if you use a plot theming function from this package and want to
save it as a raster image (JPG, PNG, etc.), use this function to save the
plot to a desired location.
}
