% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_palli.R
\name{scale-palli}
\alias{scale-palli}
\alias{scale_color_palli}
\alias{scale_fill_palli}
\title{Palli Color and Fill Scale}
\usage{
scale_color_palli(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)

scale_fill_palli(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)
}
\arguments{
\item{palette}{String, the color palette to use (defaults to \code{main}). There
are several options available: \code{main}, \code{playful}, \code{heat} (for heat meaps),
\code{heat_mono} (for monochromatic heatmaps), \code{full}, \code{decent}, \code{bright}.}

\item{discrete}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} when using a
continuous scale.}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE} changes
the order of the colors.}

\item{na.value}{String, the color for NA values.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}
for discrete scales and \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}} for continuous
scales.}
}
\description{
These color and fill functions are useful for generating the Palli color
scheme for \code{ggplot2} plots.
}
