% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_pedscience.R
\name{scale-pedscience}
\alias{scale-pedscience}
\alias{scale_color_pedscience}
\alias{scale_fill_pedscience}
\title{PedScience Color and Fill Scale}
\usage{
scale_color_pedscience(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)

scale_fill_pedscience(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)
}
\arguments{
\item{palette}{String, the color palette to use (defaults to \code{main}). There
are several options available: \code{main}, \code{playful}, \code{heat} (for heat meaps),
\code{heat_mono} (for monochromatic heatmaps), \code{full}, \code{decent}, \code{bright}.}

\item{discrete}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} when using a
continuous scale.}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE} changes
the order of the colors.}

\item{na.value}{String, the color for NA values.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}
for discrete scales and \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}} for continuous
scales.}
}
\description{
These color and fill functions are useful for generating the PedScience color
scheme for \code{ggplot2} plots.
}
\details{
With these functions one can create custom color and fill scales according to
the PedScience color scheme. Implemented are also various variations of the
\code{maine} color palette. All scales can be used for discrete scales (default)
or continuous scales (with \code{discrete = FALSE}). All scales can be reversed,
if necessary, by setting the argumet \code{reverse = TRUE}.
}
\examples{
library(ggplot2)

# Change default color scale
ggplot(palmer_penguins, aes(bill_length_mm, bill_depth_mm, color = species)) +
  geom_point() +
  scale_color_pedscience()

# Change default fill scale
ggplot(palmer_penguins, aes(species, bill_length_mm, fill = species)) +
  geom_boxplot() +
  scale_fill_pedscience()
}
