% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_dksz.R
\name{scale_color_dksz}
\alias{scale_color_dksz}
\title{Color scale using the DKSZ color scheme}
\usage{
scale_color_dksz(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey70",
  ...
)
}
\arguments{
\item{palette}{String, the color palette to use (defaults to \code{"main"}).
There are several options available: \code{"main"}, \code{"playful"},
\code{"heat"} (for heat maps), \code{"heat_mono"} (for monochromatic
heatmaps), \code{"full"}, \code{"decent"}, and \code{"bright"}.}

\item{discrete}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} when
using a continuous scale.}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE}
changes the order of the colors.}

\item{na.value}{String, the color for NA values.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}.}
}
\value{
The DKSZ color scale for ggplot2.
}
\description{
Color scale using the color scheme od the DKSZ in ggplot2.
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(x = Sepal.Length, y = Petal.Length, color = Species)) + geom_point()
p + scale_color_dksz()
}
\seealso{
\code{\link{scale_fill_dksz}}
}
