% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pedscience.R
\name{scale_color_pedscience}
\alias{scale_color_pedscience}
\title{Color scale using the PedScience color scheme}
\usage{
scale_color_pedscience(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{String, the color palette to use (defaults to \code{"main"}).
There are several options available: \code{"main"}, \code{"playful"},
\code{"heat"} (for heat maps), \code{"full"}, \code{"decent"}, and
\code{"bright"}.}

\item{discrete}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} when
using a continuous scale.}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE}
changes the order of the colors.}

\item{...}{Additional arguments passed to \code{\link{scale_color_gradientn}}.}
}
\value{
The PedScience color scale for ggplot2.
}
\description{
Color scale using the PedScience color scheme in ggplot2.
}
