% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_palli.R
\name{scale_fill_palli}
\alias{scale_fill_palli}
\title{Fill scale using the Palliativ color scheme}
\usage{
scale_fill_palli(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)
}
\arguments{
\item{palette}{String, the color palette to use (defaults to \code{"main"}).
There are several options available: \code{"main"}, \code{"playful"},
\code{"heat"} (for heat maps), \code{"full"}, \code{"decent"}, and
\code{"bright"}.}

\item{discrete}{Logical, defaults to \code{TRUE}. Set to \code{FALSE} when
using a continuous scale.}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE}
changes the order of the colors.}

\item{na.value}{String, the color for NA values.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{scale_color_gradientn}}.}
}
\value{
The Kinderpalliativzentrum fill scale.
}
\description{
Fill scale using the color scheme of the Kinderpalliativzentrum in ggplot2.
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) + geom_boxplot()
p + scale_fill_palli()
}
\seealso{
\code{\link{scale_color_palli}}
}
