% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_palli_theme.R
\name{set_palli_theme}
\alias{set_palli_theme}
\title{Set Palli Theme}
\usage{
set_palli_theme(accent = "green", palette = "main", reverse = FALSE)
}
\arguments{
\item{accent}{String, the accent color to choose (defaults to \code{green}). There
are several options available: \code{green}, \code{orange}, \code{yellow}, \code{blue},
\code{grey}}

\item{palette}{String, the color palette to use (defaults to \code{main}). There
are several options available: \code{main}, \code{playful}, \code{full}, \code{decent}, \code{bright}.}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE} changes
the order of the colors.}
}
\description{
Set the Palli color theme for ggplot2 plots. This is a more coherent
version of theming using the thematic package.
}
