% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtsummary_apa.R
\name{theme_gtsummary_apa}
\alias{theme_gtsummary_apa}
\title{APA-like gtsummary theme}
\usage{
theme_gtsummary_apa(set_theme = TRUE)
}
\arguments{
\item{set_theme}{Logial, defaults to TRUE. Indicates whether the theme should
be loaded.}
}
\value{
An APA-like gtsummary table
}
\description{
Set an APA-like theme for \code{gtsummary} tables, i.e. means and standard
deviations with two decimals are calculated for continuous variables, the
Welch test (as the default t.test function in R calculates a Welch test) is
the default statistical test for a two group comparison, and the method for
correcting p-values is the false discovery rate correction.
}
\examples{
library(gtsummary)
theme_gtsummary_apa()

trial \%>\%
 select(trt, age, response, ttdeath) \%>\%
 tbl_summary(by = trt) \%>\%
 add_n() \%>\%
 add_p() \%>\%
 add_q(quiet = TRUE)
}
\seealso{
\code{\link[=theme_gt_apa]{theme_gt_apa()}}
}
