% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{theming_colors}
\alias{theming_colors}
\title{Theming Color Values}
\format{
A list of length 4.
\describe{
\item{personal}{Personal colors}
\item{pedscience}{PedScience colors}
\item{palli}{Palliativzentrum colors}
\item{dksz}{DKSZ colors}
}
}
\usage{
theming_colors
}
\description{
This list contains all colors required for the custom theming functions.
}
\keyword{datasets}
