% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{theming_palettes}
\alias{theming_palettes}
\title{Theming Color Palettes}
\format{
A list of length 4.
\describe{
\item{personal}{Personal color palette}
\item{pedscience}{PedScience color palette}
\item{palli}{Palliativzentrum color palette}
\item{dksz}{DKSZ color palette}
}
}
\usage{
theming_palettes
}
\description{
This list contains all color palettes required for the custom theming
functions.
}
\keyword{datasets}
