% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tibble.R
\name{to_tibble}
\alias{to_tibble}
\title{Convert Matrix Like Object to Tibble}
\usage{
to_tibble(data, row_name)
}
\arguments{
\item{data}{Matrix.}

\item{row_name}{String, specifiying the column label for rownames.}
}
\value{
A tibble.
}
\description{
Convert "base R" function output  to a nicely printed tibble and get the
rownames into a dedicated column.
}
