% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_analysis_template}
\alias{use_analysis_template}
\title{Use Analysis Template}
\usage{
use_analysis_template(folder = NA)
}
\arguments{
\item{folder}{Folder to where the template should be copied.
Defaults to "03 R". The folder must be in the current working directory.}
}
\description{
Execute this function to use a default raw analysis template. By default, the
template will be placed into folder \verb{03 R}, but the destination can be
changed via \code{destination_path}
}
