% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_custom_theme.R
\name{use_carto_theme}
\alias{use_carto_theme}
\title{Set a ggplot2 Theme Using CARTO Colors}
\usage{
use_carto_theme(
  qualitative_scale = "Vivid",
  sequential_scale = "Mint",
  qualitative_accent = TRUE,
  font = "Roboto",
  reverse = FALSE
)
}
\arguments{
\item{qualitative_scale}{String, one of
\itemize{
\item "Vivid" (the default)
\item "Safe"
\item "Prism"
\item "Pastel"
\item "Bold"
\item "Antique"
}}

\item{sequential_scale}{String, for sequential scale one of
\itemize{
\item "Mint" (the default)
\item "TealGrn"
\item "Teal"
\item "SunsetDark"
\item "Sunset"
\item "RedOr"
\item "PurpOr"
\item "Purp"
\item "PinkYl"
\item "Peach"
\item "OrYel"
\item "Magenta"
\item "Emrld"
\item "DarkMint"
\item "BurgYl"
\item "Burg"
\item "BrwnYl"
\item "BluYl"
\item "BluGrn"
For diverging scales one of
\item "Tropic"
\item "Temps"
\item "TealRose"
\item "Geyser"
\item "Fall"
\item "Earth"
\item "ArmyRose"
}}

\item{qualitative_accent}{Logiocal, should the accent color come from the
qualitative scale?}

\item{font}{String, a font family from Google Fonts}

\item{reverse}{Logical, should color scale be reversed?}
}
\value{
Nothing, called for side effects only
}
\description{
https://carto.com/carto-colors/
}
\examples{
library(ggplot2)

use_carto_theme()
palmer_penguins |>
  ggplot(aes(bill_length_mm, flipper_length_mm, color = species)) +
  geom_point()


palmer_penguins |>
  ggplot(aes(bill_length_mm, flipper_length_mm, color = body_mass_g)) +
  geom_point()
}
