% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_custom_theme.R
\name{use_custom_theme}
\alias{use_custom_theme}
\title{Set a Custom ggplot2 Theme}
\usage{
use_custom_theme(
  palette = "main",
  theme = "personal",
  accent_color,
  font = "Roboto",
  reverse = FALSE
)
}
\arguments{
\item{palette}{A palette in \code{theming_palettes}. Can be one of
\itemize{
\item \code{"main"}
\item \code{"playful"}
\item \code{"full"}
\item \code{"decent"}
\item \code{"bright"}
}}

\item{theme}{Which theme to pick the palette from. Can be one of
\itemize{
\item \code{"personal"}
\item \code{"pedscience"}
\item \code{"palli"}
\item \code{"dksz"}
}}

\item{accent_color}{The accent color. Can be a string for a HEX value. The
default is the palettes accent color in \code{theming_palettes}}

\item{font}{String, a Google Font. Default is \code{"Roboto"}}

\item{reverse}{Logical. Should the palette be reversed?}
}
\value{
Nothing, called for side effects
}
\description{
Set a Custom ggplot2 Theme
}
\examples{
use_custom_theme(palette = "full")

mpg \%>\%
  ggplot(aes(displ, hwy, color = as_factor(cyl))) +
  geom_point()

# Reset the theme with
use_base_theme()
}
