% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_data_cleaning_template}
\alias{use_data_cleaning_template}
\title{Use Data Cleaning Template}
\usage{
use_data_cleaning_template(folder = NA, file_format = "csv")
}
\arguments{
\item{folder}{Folder to where the template should be copied.
Defaults to "03 R". The folder must be in the current working directory.}

\item{file_format}{Which file format do you want to clean? You can choose
between "csv" (default), "excel", and "spss"}
}
\description{
Execute this function to use a data cleaning template containing the relevant
packages to clean a file of the chosen type.
}
