% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_helper_functions}
\alias{use_helper_functions}
\title{Create a Helper Functions File}
\usage{
use_helper_functions(folder = NA)
}
\arguments{
\item{folder}{Folder to where the template should be copied.
Defaults to "03 R". The folder must be in the current working directory.}
}
\description{
Execute this function to generate a helper-functions.R file in the \verb{03 R}
folder, but the destination can be changed via \code{folder}.
}
