% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_latex_template}
\alias{use_latex_template}
\title{Use A LaTeX Document Template}
\usage{
use_latex_template(folder = NA)
}
\arguments{
\item{folder}{Folder to where the template should be copied.
Defaults to "03 R". The folder must be in the current working directory.}
}
\description{
Execute this function to generate a LaTeX document template with prepopulated
preamble to get your hands on quickly. By default, the template will be
placed into folder \verb{05 Reports}, but the destination can be changed via
\code{folder}.
}
