% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pedscience_theme.R
\name{use_pedscience_theme}
\alias{use_pedscience_theme}
\title{Set PedScience Theme}
\usage{
use_pedscience_theme(
  accent = "blue",
  palette = "main",
  font = "Roboto",
  font_scale = 1,
  reverse = FALSE
)
}
\arguments{
\item{accent}{String, the accent color to choose (defaults to \code{blue}). There
are several options available: \code{blue}, \code{coral}, \code{baby_blue}, \code{dark_grey},
\code{light_grey}, \code{light_blue}}

\item{palette}{String, the color palette to use (defaults to \code{main}). There
are several options available: \code{main}, \code{playful}, \code{full}, \code{decent}, \code{bright}.}

\item{font}{String, a font of choice (defaults to Roboto)}

\item{font_scale}{Integer, scaling factor for fonts (defaults to 1)}

\item{reverse}{Logical, defaults to \code{FALSE}. Setting this to \code{TRUE} changes
the order of the colors.}
}
\description{
Set the PedScience color theme for ggplot2 plots. This is a more coherent
version of theming using the thematic package.
}
\examples{
library(ggplot2)

# Ordinary ggplot2 plot
test_plot <- palmer_penguins \%>\%
  ggplot(aes(bill_length_mm, flipper_length_mm)) +
  geom_smooth(method = "loess") +
  geom_point(aes(color = species))
test_plot

# Setting the theme
set_pedscience_theme()

test_plot
}
