% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_summaries.R
\name{win_mean_se}
\alias{win_mean_se}
\title{Compute Winsorized Mean and its Standard Error for ggplot2}
\usage{
win_mean_se(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A tibble with columns \code{ymin}, \code{mean}, and \code{ymax}.
}
\description{
Compute 20\% winsorized mean and standard error for a given vector. Can be used
as \code{fun.data} in ggplot2's \code{\link[ggplot2]{stat_summary}}.
}
\examples{
x <- rnorm(200)

win_mean_se(x)
}
