% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bene.R
\name{theme_bene}
\alias{theme_bene}
\title{Custom Theme for ggplot2}
\usage{
theme_bene(base_size = 11)
}
\arguments{
\item{base_size}{Integer, the text base size}
}
\value{
Theme for ggplot2.
}
\description{
Custom \code{ggplot2} theme based on \code{theme_light()} but using Google's
Roboto as default font. To use the font, the package \code{extrafont} must be
loaded.
}
\examples{
\dontrun{
library(ggplot2)
library(extrafont)

ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  theme_bene()
}
}
