% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_project.R
\name{use_custom_folder}
\alias{use_custom_folder}
\title{Create A Custom Folder}
\usage{
use_custom_folder(path = NA, folder_name, add = FALSE)
}
\arguments{
\item{path}{Where should the folder be created?}

\item{folder_name}{What should the folder be called? This must be a string.}

\item{add}{Should the folder be added even if the name already exists?
Defaults to \code{FALSE}.}
}
\description{
Create a new folder with formatted number and custom name in the current
working directory. The newly created folder will be labelled with the number
n + 1 with n being the number of folders already existent in the working
directory. If a folder with the same name (must not be the same number)
already exists, you can choose to add another one with the same name (and
different number).
}
